use proper awk syntax in functions

+++ git-undo.awk
@@ -13,13 +13,14 @@ BEGIN {
 
 
 function getFileNames(cmd) {
-	var parts = cmd.split(" ")
-	for(var p = 0; p < parts.length; p++){
-		if(parts[p].indexOf("-") === 0){
-			parts[p] = ""
+	nfiles = split(cmd, parts, / /)
+	for (i=1; i<=nfiles; i++) {
+		if (!match(parts[i], /^-/)) {
+			files += " "
+			files += parts[i]
 		}
 	}
-	return parts
+	return substr(files, 1)
 }
 
 {
@@ -32,35 +33,33 @@ function getFileNames(cmd) {
 	undoCommand()
 	if (info) {
 		print info
+		if (command) {
+	#		if ((program.rawArgs.indexOf("-f") > -1 || program.rawArgs.indexOf("--fix") > -1) && autorun) {
+	#			print "Running " command
+	#			exec(command, function(){
+	#				print "Completed"
+	#			})
+	#		} else {
+				print command
+	#		}
+		} else if (autorun) {
+			print "No undo command necessary"
+		} else {
+			print "No undo command known"
+		}
+		foundGit = 1
 	} else {
 		print "I didn't recognize that command"
-		return
 	}
-	if (command) {
-#		if ((program.rawArgs.indexOf("-f") > -1 || program.rawArgs.indexOf("--fix") > -1) && autorun) {
-#			print "Running " command
-#			exec(command, function(){
-#				print "Completed"
-#			})
-#		} else {
-			print command
-#		}
-	} else if (autorun) {
-		print "No undo command necessary"
-	} else {
-		print "No undo command known"
-	}
-	foundGit = 1
-	break
 }
 
 function undoCommand() {
-	/git init/ {
+	if (/git init/) {
 		info = "This created a .git folder in the current directory. You can remove it."
 		undo = "rm -rf .git"
 		autorun = 1
 	}
-	/git clone/ {
+	if (/git clone/) {
 		var outputfolder = null
 		var cloned_into = cmd.split("git clone ")[1].split(" ")
 		if(cloned_into.length > 1){
@@ -84,8 +83,8 @@ function undoCommand() {
 			autorun = 0
 		}
 	}
-	/git add/ {
-		var filenames = getFileNames(cmd.split("git add ")[1])
+	if (/git add/) {
+		var filenames = getFileNames(substr($0, 9)) # remove "git add " prefix
 		info = "This added files to the changes staged for commit. All changes to files will be removed from staging for this commit, but remain saved in the local file system."
 		if(filenames.indexOf(".") > -1 || filenames.indexOf("*") > -1){
 			info += "\nUsing . or * affects all files, so you will need to run 'git reset <file>' on each file you didn't want to add."
@@ -96,7 +95,7 @@ function undoCommand() {
 			autorun = 1
 		}
 	}
-	/git rm/ {
+	if (/git rm/) {
 		filenames = cmd.split("git rm ")[1]
 		if(/--cached/){
 			info = "This took files out of the changes staged for commit. All changes will be re-added to staging for this commit."
@@ -108,19 +107,19 @@ function undoCommand() {
 		}
 		autorun = 1
 	}
-	/git mv/ {
+	if (/git mv/) {
 		var old_name = cmd.split("git mv ")[1].split(" ")[0]
 		var new_name = cmd.split("git mv ")[1].split(" ")[1]
 		info = "This moved the file (named " + old_name + ") to " + new_name + ". It can be moved back."
 		undo = "git mv " + new_name + " " + old_name
 		autorun = 1
 	}
-	/git checkout/ {
+	if (/git checkout/) {
 		info = "git checkout moved you into a different branch of the repo. You can checkout any branch by name, or checkout the last one using -"
 		undo = "git checkout -"
 		autorun = 1
 	}
-	/git remote add/ {
+	if (/git remote add/) {
 		var repo_name = cmd.split("git remote add ")[1].split(" ")[0]
 		var repo_url = cmd.split("git remote add ")[1].split(" ")[1]
 
@@ -129,14 +128,14 @@ function undoCommand() {
 		undo = "git remote rm " + repo_name
 		autorun = 1
 	}
-	/git remote remove/ || /git remote rm/ {
+	if (/git remote remove/ || /git remote rm/) {
 		var repo_name = cmd.split("git remote ")[1].split(" ")[1]
 
 		info = "This removed a remote repo (named " + repo_name + ")"
 		info += "\nIt needs to be added back using git remote add " + repo_name + " <git-url>"
 		autorun = 0
 	}
-	/git remote set-url/ {
+	if (/git remote set-url/) {
 		var repo_name = cmd.split("git remote set-url ")[1].split(" ")[0]
 		var repo_url = cmd.split("git remote set-url ")[1].split(" ")[1]
 
@@ -144,32 +143,32 @@ function undoCommand() {
 		info += "\nIt can be removed (using git remote rm) or set again (using git remote set-url)."
 		autorun = 0
 	}
-	/git remote rename/ {
+	if (/git remote rename/) {
 		var old_name = cmd.split("git remote rename ")[1].split(" ")[0]
 		var new_name = cmd.split("git remote rename ")[1].split(" ")[1]
 		info = "This changed the remote repo (named " + old_name + ") to have the name " + new_name + ". It can be reset."
 		undo = "git remote rename " + new_name + " " + old_name
 		autorun = 1
 	}
-	/git commit/ {
+	if (/git commit/) {
 		info = "This saved your staged changes as a commit, which can be updated with git commit --amend or completely uncommited:"
 		undo = "git reset --soft HEAD^"
 	}
-	/git revert/ {
+	if (/git revert/) {
 		info = "This made a new commit to retract a commit. You can undo *the revert commit* using a more extreme approach:"
 		undo = "git reset --soft HEAD^"
 	}
-	/git fetch/ {
+	if (/git fetch/) {
 		info = "This updated the local copy of all branches in this repo. Un-updating master (and you can do other branches, too)."
 		undo = "git update-ref refs/remotes/origin/master refs/remotes/origin/master@{1}"
 		autorun = 1
 	}
-	/git pull/ || /git merge/ {
+	if (/git pull/ || /git merge/) {
 		info = "This merged another branch (local or remote) into your current branch. This resets you to the last version."
 		undo = "git reset --hard HEAD^"
 		autorun = 1
 	}
-	/git push/ {
+	if (/git push/) {
 		autorun = 0
 		info = "This uploaded all of your committed changes to a remote repo. It may be difficult to reverse it."
 		info += "\nYou can use git revert <commit_id> to tell repos to turn back these commits."
@@ -178,14 +177,14 @@ function undoCommand() {
 			info += "\nIf you are hosting this app on Heroku, run 'heroku rollback' to reset your app now."; 
 		}
 	}
-	/git branch/ {
+	if (/git branch/) {
 		autorun = 1
 		if(/ -D/){
 			# delete branch
 			info = "You deleted a branch. You can use 'git branch' to create it again, or 'git pull' to restore it from a remote repo."
 			autorun = 0
 		}
-		/git branch / {
+		if (/git branch /) {
 			# create branch
 			var branch_name = cmd.split("git branch ")[1].split(" ")[0]
 			if(branch_name.length && branch_name[0] != "-"){
@@ -198,7 +197,7 @@ function undoCommand() {
 			info = "git branch on its own doesn't change the repo; it just lists all branches. Use it often!"
 		}
 	}
-	/git stash/ {
+	if (/git stash/) {
 		if(/stash list/){
 			info = "git stash list doesn't change the repo; it just tells you the stashed changes which you can restore using git stash apply."
 			autorun = 1
@@ -213,40 +212,40 @@ function undoCommand() {
 			autorun = 1
 		}
 	}
-	/git archive/ {
+	if (/git archive/) {
 		info = "This created an archive of part of the repo - you can delete it using 'rm -rf <archive_file_or_folder>'."
 		autorun = 0
 	}
 	# harmless
-	/git cat-file/ {
+	if (/git cat-file/) {
 		info = "git cat-file doesn't change the repo; it just tells you the type of an object in the repo."
 		autorun = 1
 	}
-	/git diff/ {
+	if (/git diff/) {
 		info = "git diff doesn't change the repo; it just tells you the changes waiting for commit OR the changes between branches. Use it often!"
 		autorun = 1
 	}
-	/git grep/ {
+	if (/git grep/) {
 		info = "git grep doesn't change the repo; it's a search tool. Use grep and git grep often!"
 		autorun = 1
 	}
-	/git ls-tree/ {
+	if (/git ls-tree/) {
 		info = "git ls-tree doesn't change the repo; it just tells you about an object in the git repo."
 		autorun = 1
 	}
-	/git show/ {
+	if (/git show/) {
 		info = "git show doesn't change the repo; it just tells you the changes waiting for commit OR the changes between branches."
 		autorun = 1
 	}
-	/git log/ {
+	if (/git log/) {
 		info = "git log doesn't change the repo; it just lists the last several commits in this branch. Use it often!"
 		autorun = 1
 	}
-	/git status/ {
+	if (/git status/) {
 		info = "git status doesn't change the repo; it just tells you what changes there are. Use it often!"
 		autorun = 1
 	}
-	/git remote/ {
+	if (/git remote/) {
 		info = "git remote (without additional arguments) doesn't change the repo; it just tells you what remotes there are. Use it often!"
 		autorun = 1
 	}
